deliverable3 done for steph yo

+++ iOSObjectiveCXXSwift/MangoScene.mm
@@ -12,27 +12,33 @@
 
 -(id)init{
     self = [super init];
-    ENV = (environment *)createMangoEnvironment();
     return self;
 }
 
 -(void)loadSceneWithName:(NSString *)name{
+    if(ENV != NULL)
+        deleteEnv(ENV);
     NSString* path = [[NSBundle mainBundle] pathForResource:name
                                                      ofType:@"lisp"];
     NSString* content = [NSString stringWithContentsOfFile:path
                                                   encoding:NSUTF8StringEncoding
                                                      error:NULL];
+    ENV = (environment *)createMangoEnvironment();
     [self runString:content];
     [self triggerEvent:@"SCENE_LOAD"];
 }
 
 -(void)runString:(NSString *)program{
+    if(ENV == NULL)
+        return;
     expression * prog = parseList((char*)[program cStringUsingEncoding:NSASCIIStringEncoding], [program length]);
     runProgram(prog, ENV);
     deleteExpression(prog);
 }
 
 -(void)triggerEvent:(NSString *)event{
+    if(ENV == NULL)
+        return;
     char * cevent = (char *)[event cStringUsingEncoding:NSASCIIStringEncoding];
     environmentIterator it = ENV->find(cevent);
     if(it != ENV->end()){
